Hash Table Array
HashTable in Java is a type of Collections framework which is used to store and manipulate the group of objects. It is an array of the list in java (known as a bucket) which is used to implement a hash table, which maps keys to values. Hashtable contains unique elements